refactor: TeamOccupancyRate를 도메인의 위치 변경
