+++ android/app/src/main/java/com/yagubogu/data/dto/response/StatsStadiumOccupancyRateResponse.kt
@@ -1,13 +1,13 @@
 package com.yagubogu.data.dto.response
 
-import com.yagubogu.domain.model.TeamOccupancyRate
+import com.yagubogu.presentation.stats.stadium.TeamOccupancyRate
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class StatsStadiumOccupancyRateResponse(
     @SerialName("teams")
-    val teams: List<TeamDto>, // 팀 정보(팀 ID, 팀 이름, 점유율)
+    val teams: List<TeamOccupancyRateDto>, // 팀 정보(팀 ID, 팀 이름, 점유율)
 ) {
     fun toDomains(): List<TeamOccupancyRate> = teams.map { it.toDomain() }
 }
