+++ android/app/src/main/java/com/yagubogu/presentation/stats/stadium/TeamOccupancyRate.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.domain.model
+package com.yagubogu.presentation.stats.stadium
 
 data class TeamOccupancyRate(
     val id: Long,
